split monoids
