+++ src/data-structure/segment-tree/presets/monoid/operator-structure-add.hpp
@@ -1,0 +1,20 @@
+#pragma once
+
+namespace luz::monoid {
+
+  template < typename T >
+  class RangeAddQueryMonoid {
+
+   public:
+    using value_type = T;
+
+    static constexpr T operation(T a, T b) {
+      return a + b;
+    }
+
+    static constexpr T identity() {
+      return T();
+    }
+  };
+
+} // namespace luz::monoid
