+++ src/data-structure/segment-tree/presets/range-update-query-solver.hpp
@@ -1,27 +1,7 @@
 #pragma once
 
 #include "src/data-structure/segment-tree/range-mapping-point-fold-segment-tree.hpp"
-
-#include <algorithm>
-
-namespace luz::monoid {
-
-  template < typename T, T ID >
-  class RangeUpdateQueryMonoid {
-
-   public:
-    using value_type = T;
-
-    static constexpr T operation(T a, T b) {
-      return b == ID ? a : b;
-    }
-
-    static constexpr T identity() {
-      return ID;
-    }
-  };
-
-} // namespace luz::monoid
+#include "src/data-structure/segment-tree/presets/monoid/operator-structure-update.hpp"
 
 namespace luz {
